refactor: move evaluator in seperate folder
