+++ src/intelligence_layer/evaluation/run_evaluation.py
@@ -16,7 +16,7 @@ from intelligence_layer.evaluation.aggregation.file_aggregation_repository impor
 from intelligence_layer.evaluation.dataset.file_dataset_repository import (
     FileDatasetRepository,
 )
-from intelligence_layer.evaluation.evaluation.evaluator import Evaluator
+from intelligence_layer.evaluation.evaluation.evaluator.evaluator import Evaluator
 from intelligence_layer.evaluation.evaluation.file_evaluation_repository import (
     FileEvaluationRepository,
 )
