+++ tests/evaluation/test_async_evaluation_repository.py
@@ -5,12 +5,12 @@ from uuid import uuid4
 from pytest import FixtureRequest, fixture, mark
 
 from intelligence_layer.core.tracer.tracer import utc_now
-from intelligence_layer.evaluation.evaluation.async_evaluation import (
-    AsyncEvaluationRepository,
-)
 from intelligence_layer.evaluation.evaluation.domain import (
     EvaluationOverview,
     PartialEvaluationOverview,
+)
+from intelligence_layer.evaluation.evaluation.evaluator.async_evaluator import (
+    AsyncEvaluationRepository,
 )
 from intelligence_layer.evaluation.evaluation.file_evaluation_repository import (
     AsyncFileEvaluationRepository,
